Refactor WebhookFlowHandler into config_entry_flow
