+++ homeassistant/components/ifttt/__init__.py
@@ -12,7 +12,7 @@ import voluptuous as vol
 
 import homeassistant.helpers.config_validation as cv
 from homeassistant.const import CONF_WEBHOOK_ID
-from homeassistant.helpers import webhook_config_entry_flow
+from homeassistant.helpers import config_entry_flow
 
 REQUIREMENTS = ['pyfttt==0.3']
 DEPENDENCIES = ['webhook']
@@ -97,7 +97,7 @@ async def async_unload_entry(hass, entry):
     hass.components.webhook.async_unregister(entry.data[CONF_WEBHOOK_ID])
     return True
 
-webhook_config_entry_flow.register_webhook_flow(
+config_entry_flow.register_webhook_flow(
     DOMAIN,
     'IFTTT Webhook',
     {
